Fix all other test compilation errors due to untyped literals being put into Ptr

+++ rest/access_test.go
@@ -285,7 +285,7 @@ func TestUserHasDocAccessDocNotFound(t *testing.T) {
 			QueryPaginationLimit: base.Ptr(2),
 			CacheConfig: &CacheConfig{
 				RevCacheConfig: &RevCacheConfig{
-					MaxItemCount: base.Ptr(0),
+					MaxItemCount: base.Ptr(uint32(0)),
 				},
 				ChannelCacheConfig: &ChannelCacheConfig{
 					MaxNumber: base.Ptr(0),
